refactor(poseidon2): rename bb31 to baby_bear
